+++ test/integration/DecimalsTests.t.sol
@@ -122,12 +122,12 @@ contract DecimalsTests is BaseTest {
 
     function test_fuzz_node_swing_price_deposit_never_exceeds_max_6decimals(
         uint64 maxSwingFactor,
-        uint256 targetReserveRatio,
+        uint64 targetReserveRatio,
         uint256 seedAmount,
         uint256 depositAmount
     ) public {
         maxSwingFactor = uint64(bound(maxSwingFactor, 0.01 ether, 0.99 ether));
-        targetReserveRatio = bound(targetReserveRatio, 0.01 ether, 0.99 ether);
+        targetReserveRatio = uint64(bound(targetReserveRatio, 0.01 ether, 0.99 ether));
         seedAmount = bound(seedAmount, 1 ether, maxDeposit);
         depositAmount = bound(depositAmount, 1 ether, maxDeposit);
 
@@ -183,12 +183,12 @@ contract DecimalsTests is BaseTest {
 
     function test_fuzz_node_swing_price_redeem_never_exceeds_max_6decimals(
         uint64 maxSwingFactor,
-        uint256 targetReserveRatio,
+        uint64 targetReserveRatio,
         uint256 seedAmount,
         uint256 withdrawalAmount
     ) public {
         maxSwingFactor = uint64(bound(maxSwingFactor, 0.01 ether, 0.99 ether));
-        targetReserveRatio = bound(targetReserveRatio, 0.01 ether, 0.99 ether);
+        targetReserveRatio = uint64(bound(targetReserveRatio, 0.01 ether, 0.99 ether));
         seedAmount = bound(seedAmount, 100 ether, 1e36);
 
         deal(address(testToken6), address(user), seedAmount);
